Fix the tests as per changes in hooks and trigger
